+++ src/main/java/com/spring/users/repository/PersonRepository.java
@@ -3,6 +3,6 @@ package com.spring.users.repository;
 import com.spring.users.model.Person;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PersonRepository extends CrudRepository<Person, Long> {
+public interface PersonRepository extends CrudRepository<Person, Integer> {
 
 }
