+++ src/main/java/com/spring/users/repository/UsersRepository.java
@@ -3,6 +3,6 @@ package com.spring.users.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.spring.users.model.Users;
 
-public interface UsersRepository extends CrudRepository<Users, Long>{
+public interface UsersRepository extends CrudRepository<Users, Integer>{
 
 }
